refactor: usign signature and usdc for bets
